+++ core/services/job/orm.go
@@ -47,7 +47,7 @@ type ORM interface {
 	RecordError(jobID int32, description string, qopts ...pg.QOpt) error
 	// TryRecordError is a helper which calls RecordError and logs the returned error if present.
 	TryRecordError(jobID int32, description string, qopts ...pg.QOpt)
-	DismissError(ctx context.Context, errorID int32) error
+	DismissError(ctx context.Context, errorID int64) error
 	FindSpecError(id int64, qopts ...pg.QOpt) (SpecError, error)
 	Close() error
 	PipelineRuns(jobID *int32, offset, size int) ([]pipeline.Run, int, error)
@@ -334,7 +334,7 @@ func (o *orm) TryRecordError(jobID int32, description string, qopts ...pg.QOpt) 
 	o.lggr.ErrorIf(err, fmt.Sprintf("Error creating SpecError %v", description))
 }
 
-func (o *orm) DismissError(ctx context.Context, ID int32) error {
+func (o *orm) DismissError(ctx context.Context, ID int64) error {
 	q := o.q.WithOpts(pg.WithParentCtx(ctx))
 	res, cancel, err := q.ExecQIter("DELETE FROM job_spec_errors WHERE id = $1", ID)
 	defer cancel()
